txsub package reorg and cleanup
